requirements.txt
PyQt6>=6.9.0
pyserial>=3.5
pyserial-asyncio>=0.6
